ES-1638 (#359)
* ES-1638


* ES-1638


* ES-1638


* ES-1638


---------

+++ signup-ui/configure_start.sh
@@ -5,39 +5,44 @@ set -e
 
 echo "Downloading pre-requisites started."
 
-#i18n bundle
-echo "Downloading i18n bundle files"
-wget --no-check-certificate --no-cache --no-cookies $artifactory_url_env/artifactory/libs-release-local/i18n/esignet-signup-i18n-bundle.zip -O $i18n_path/esignet-signup-i18n-bundle.zip
+# Check if $i18n_url_env is not empty
+if [[ -n "$i18n_url_env" ]]; then
+    echo "i18n_url_env is set: $i18n_url_env"
+    wget --no-check-certificate --no-cache --no-cookies $i18n_url_env -O $i18n_path/esignet-signup-i18n-bundle.zip
 
-echo "unzip i18n bundle files.."
-chmod 775 $i18n_path/*
+    echo "unzip i18n bundle files.."
+    chmod 775 $i18n_path/*
+    cd $i18n_path
+    unzip -o esignet-signup-i18n-bundle.zip
+    rm esignet-signup-i18n-bundle.zip
+    echo "unzip i18n bundle completed."
+fi
 
-cd $i18n_path
-unzip -o esignet-signup-i18n-bundle.zip
-rm esignet-signup-i18n-bundle.zip
+# Check if $theme_url_env is not empty
+if [[ -n "$theme_url_env" ]]; then
+    echo "theme_url_env is set: $theme_url_env"
+    wget --no-check-certificate --no-cache --no-cookies $theme_url_env -O $theme_path/esignet-signup-theme.zip
 
-#theme bundle
-echo "Downloading theme bundle files"
-wget --no-check-certificate --no-cache --no-cookies $artifactory_url_env/artifactory/libs-release-local/theme/esignet-signup-theme.zip -O $theme_path/esignet-signup-theme.zip
+    echo "unzip theme files.."
+    chmod 775 $theme_path/*
+    cd $theme_path
+    unzip -o esignet-signup-theme.zip
+    rm esignet-signup-theme.zip
+    echo "unzip theme completed."
+fi
 
-echo "unzip i18n bundle files.."
-chmod 775 $theme_path/*
+# Check if $images_url_env is not empty
+if [[ -n "$images_url_env" ]]; then
+    echo "images_url_env is set: $images_url_env"
+    wget --no-check-certificate --no-cache --no-cookies $images_url_env -O $image_path/esignet-signup-image.zip
 
-cd $theme_path
-unzip -o esignet-signup-theme.zip
-rm esignet-signup-theme.zip
-
-#image bundle
-echo "Downloading image bundle files"
-wget --no-check-certificate --no-cache --no-cookies $artifactory_url_env/artifactory/libs-release-local/image/esignet-signup-image.zip -O $image_path/esignet-signup-image.zip
-
-echo "unzip image bundle files.."
-chmod 775 $image_path/*
-
-cd $image_path
-unzip -o esignet-signup-image.zip
-rm esignet-signup-image.zip
-
+    echo "unzip image files.."
+    chmod 775 $image_path/*
+    cd $image_path
+    unzip -o esignet-signup-image.zip
+    rm esignet-signup-image.zip
+    echo "unzip image completed."
+fi
 
 echo "Pre-requisites download completed."
